+++ src/components/bookmark -button/bookmark -button.tsx
@@ -1,22 +1,32 @@
 import { FavoritsButtonClassName } from '../../const';
+import { useAuthorizationStatus } from '../../hooks/use-authorization-status';
+// import { useFavorites } from '../../hooks/use-favorite';
+// import { useAppDispatch } from '../../hooks';
+// import { useEffect } from 'react';
+// import { fetchFavoritesAction } from '../../store/api-actions';
+import { useNavigate } from 'react-router-dom';
+import { AppRoute } from '../../const';
+import { AuthorizationStatus } from '../../const';
+import { useAppDispatch, useAppSelector } from '../../hooks';
+import { toggleFavoriteAction } from '../../store/api-actions';
+
 
 type BookmarkButtonProps = {
-  isFavorite: boolean;
-  handleFavoriteClick: () => void;
+  idFavorite:string;
+  // isFavorite: boolean;
+  // handleFavoriteClick: () => void;
   buttonClassName: FavoritsButtonClassName;
 };
 
 function BookmarkButton({
-  isFavorite,
-  handleFavoriteClick,
+  // isFavorite,
+  idFavorite,
+  // handleFavoriteClick,
   buttonClassName,
 }: BookmarkButtonProps): JSX.Element {
   const buttonClass = `${buttonClassName}__bookmark-button`;
   const iconClass = `${buttonClassName}__bookmark-icon`;
 
-  // const iconSizes = buttonClassName === FavoritsButtonClassName.Offer
-  // ? { width: 31, height: 33 }
-  // : { width: 18, height: 19 };
 
   const BUTTON_SIZES = {
     'place-card__bookmark-icon': { width: 18, height: 19 },
@@ -25,6 +35,27 @@ function BookmarkButton({
   const { width, height } = BUTTON_SIZES[
     iconClass as keyof typeof BUTTON_SIZES
   ] || { width: 18, height: 19 };
+
+  const favorites = useAppSelector((state) => state.favorites);
+
+  const isFavorite = favorites.findIndex((offer)=>offer.id === idFavorite) !== -1;
+
+  const navigate = useNavigate();
+
+  const authorizationStatus = useAuthorizationStatus();
+
+
+  const dispatch = useAppDispatch();
+
+  const handleFavoriteClick = () => {
+    if (authorizationStatus !== AuthorizationStatus.Auth) {
+      navigate(AppRoute.Login);
+      return;
+    }
+
+    dispatch(toggleFavoriteAction({ id: idFavorite, isFavorite }));
+  };
+
 
   return (
     <button
